+++ src/lib.rs
@@ -7,26 +7,21 @@ mod batch_builder;
 mod client;
 mod content_builder;
 mod embed_builder;
-mod error;
 mod models;
 mod tools;
 
-pub use batch::Batch;
+pub use batch::{Batch, BatchStatus, Error as BatchError};
 pub use batch_builder::BatchBuilder;
-pub use client::Gemini;
+pub use client::{Error as ClientError, Gemini, Model};
 pub use content_builder::ContentBuilder;
-pub use error::Error;
 pub use models::{
     BatchConfig, BatchGenerateContentResponseItem, BatchMetadata, BatchOperationResponse,
-    BatchRequestItem, BatchResultItem, BatchState, BatchStats, BatchStatus, Blob, Candidate,
-    CitationMetadata, Content, FunctionCallingMode, GenerateContentRequest, GenerationConfig,
-    GenerationResponse, InlinedResponses, InputConfig, Message, MultiSpeakerVoiceConfig,
-    OutputConfig, Part, PrebuiltVoiceConfig, PromptTokenDetails, RequestMetadata,
-    RequestsContainer, Role, SafetyRating, SpeakerVoiceConfig, SpeechConfig, TaskType,
-    ThinkingConfig, UsageMetadata, VoiceConfig,
+    BatchRequestItem, BatchResultItem, BatchState, BatchStats, Blob, Candidate, CitationMetadata,
+    Content, FunctionCallingMode, GenerateContentRequest, GenerationConfig, GenerationResponse,
+    InlinedResponses, InputConfig, Message, MultiSpeakerVoiceConfig, OutputConfig, Part,
+    PrebuiltVoiceConfig, PromptTokenDetails, RequestMetadata, RequestsContainer, Role,
+    SafetyRating, SpeakerVoiceConfig, SpeechConfig, TaskType, ThinkingConfig, UsageMetadata,
+    VoiceConfig,
 };
 
 pub use tools::{FunctionCall, FunctionDeclaration, FunctionParameters, PropertyDetails, Tool};
-
-/// Result type for this crate
-pub type Result<T> = std::result::Result<T, Error>;